tp_api/pyproject.toml
[tool.poetry]
name = "tp-api"
version = "0.1.0"
description = ""
authors = ["DaiTensa <dai.tensaout@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
